Adding new code

+++ mvc-app/src/main/java/com/tribepay/mvcapp/web/BankDetailsController.java
@@ -1,8 +1,14 @@
 package com.tribepay.mvcapp.web;
 
 
+import com.tribepay.mvcapp.entity.CustomerAccountDetail;
 import com.tribepay.mvcapp.entity.User;
+import com.tribepay.mvcapp.exception.CustomUserNotFoundException;
+import com.tribepay.mvcapp.service.SearchFormImpl;
+import com.tribepay.mvcapp.service.SearchImpl;
 import com.tribepay.mvcapp.service.UserServiceImpl;
+import com.tribepay.mvcapp.springdata.CustomerAccountDetRepo;
+import com.tribepay.mvcapp.springdata.UserSpringDataRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class BankDetailsController {
@@ -20,7 +27,10 @@ public class BankDetailsController {
 
     @Autowired
     UserServiceImpl userServiceimpl;
-
+    UserSpringDataRepository repository;
+    SearchFormImpl searchForm;
+    SearchImpl searchimpl;
+    CustomerAccountDetRepo customerAccountDetRepo;
 
     private Logger logger = LoggerFactory.getLogger(this.getClass());
 
@@ -32,6 +42,15 @@ public class BankDetailsController {
     @RequestMapping(value ="/users/user" , method= RequestMethod.GET)
     public List<User> fetchAllUsers(){
       return userServiceimpl.findAllUsers();
+    }
+
+
+
+    @RequestMapping(value ="/search/{customerId}" , method= RequestMethod.GET)
+    public Optional <CustomerAccountDetail> search(@PathVariable Long customerId){
+        Optional<CustomerAccountDetail> result = customerAccountDetRepo.findByCustomerId(customerId);
+        //logger.info("retrived data"{}, customer);
+        return result;
     }
 
     @RequestMapping(path= "/users", consumes = "application/json",
@@ -59,6 +78,10 @@ public class BankDetailsController {
     }
     @RequestMapping(path="/users/del/{id}",method = RequestMethod.GET)
     public ResponseEntity<Void>deleteUser(@PathVariable int id){
+
+        if (!repository.existsById(id))
+            throw new CustomUserNotFoundException("id" + id);
+
         HttpHeaders headers = new HttpHeaders();
         userServiceimpl.deleteUserById(id);
         return new ResponseEntity<Void>(headers, HttpStatus.ACCEPTED);
